Добавил  парсер с xml в json

+++ src/main/java/Employee.java
@@ -8,7 +8,7 @@ public class Employee {
     public Employee() {
     }
 
-    public Employee(long id, String firstName, String lastName, String country, int age) {
+    public Employee(int id, String firstName, String lastName, String country, int age) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
